+++ app/initializers/ember-orbit-config.js
@@ -1,0 +1,4 @@
+export {
+  default,
+  initialize
+} from 'ember-orbit/initializers/ember-orbit-config';
