+++ app/initializers/ember-orbit-services.js
@@ -1,0 +1,4 @@
+export {
+  default,
+  initialize
+} from 'ember-orbit/initializers/ember-orbit-services';
