--- app/initializers/ember-orbit.js
@@ -1,1 +1,0 @@
-export { default, initialize } from 'ember-orbit/initializers/ember-orbit';
